+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/png" href="/dsp.png" />
-    <link href="/src/index.css" rel="stylesheet">
+    <link href="/index.css" rel="stylesheet">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>DataStack Pro</title>
   </head>
