+++ src/frontend/services/practiceModes/ExamSimulationMode.jsx
