+++ src/frontend/services/practiceModes/SurvivalMode.jsx
