+++ src/frontend/services/practicePage.js
