+++ src/frontend/utils/common.jsx
