+++ temp/normalize.py
